% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Take raw Latin Hypercube samples and rescale}
\usage{
rescale(ranges, vals)
}
\arguments{
\item{ranges}{List of mean, sd for each parameter}

\item{vals}{Matrix of samples from Latin hypercube uniform distribution.
Dimensions sample by parameter}
}
\value{
Mean, upper and lower limit adjusted
}
\description{
Take raw Latin Hypercube samples and rescale
}
