% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_get_results_dopar.R
\name{test_get_results_dopar}
\alias{test_get_results_dopar}
\title{Parallelised version}
\usage{
test_get_results_dopar(init_points, indx_in, indx_out)
}
\description{
see
https://stackoverflow.com/questions/58823105/use-doparallel-with-rcpp-function-on-window-inside-an-r-package
}
